pyproject.toml
[project]
name = "knowledge-pipeline-poc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "pawarison", email = "pawarison.ty@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "docling>=2.54.0",
    "google-generativeai>=0.8.5",
    "ipykernel>=6.30.1",
    "langchain-text-splitters>=0.3.11",
    "markdown-it-py>=4.0.0",
    "mistralai>=1.9.10",
    "numpy>=2.3.3",
    "pandas>=2.3.2",
    "python-dotenv>=1.0.1",
    "pypdf>=4.3.1",
    "pythainlp>=5.1.2",
    "tiktoken>=0.11.0",
    "python-crfsuite>=0.9.11",
]

[project.scripts]
knowledge-pipeline-poc = "knowledge_pipeline_poc:main"

[build-system]
requires = ["uv_build>=0.8.22,<0.9.0"]
build-backend = "uv_build"
